+++ libs/f-flow/src/f-draggable/is-drag-blocker.ts
@@ -1,5 +1,5 @@
 import { isClosestElementHasClass } from '@foblex/utils';
 
-export function isDragBlocker(element: HTMLElement): boolean {
+export function isDragBlocker(element: HTMLElement | SVGElement): boolean {
   return isClosestElementHasClass(element, '.f-drag-blocker');
 }